+++ backend/src/api/endpoints/v1_api_playlist.rs
@@ -29,6 +29,8 @@ fn create_config_input_for_m3u(url: &str) -> ConfigInput {
             xtream_skip_series: false,
             xtream_live_stream_without_extension: false,
             xtream_live_stream_use_prefix: true,
+            resolve_tmdb: false,
+            analyze_stream: false,
         }),
         ..Default::default()
     }
@@ -49,6 +51,8 @@ fn create_config_input_for_xtream(username: &str, password: &str, host: &str) ->
             xtream_skip_series: false,
             xtream_live_stream_without_extension: false,
             xtream_live_stream_use_prefix: true,
+            resolve_tmdb: false,
+            analyze_stream: false,
         }),
         ..Default::default()
     }